# esg-kg-platform/backend/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
...
			"Pick_Implementation.iri-or-label-or-language-or-filePath-or-functionName-or-returnType-or-validation_": {
				"properties": {
					"iri": {
						"type": "string"
					},
					"label": {
						"type": "string"
					},
					"language": {
						"type": "string"
					},
					"filePath": {
						"type": "string"
					},
					"functionName": {
						"type": "string"
					},
					"returnType": {
						"type": "string"
					},
					"validation": {
						"type": "string"
					}
				},
				"required": [
					"iri"
				],
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
			"ImplementationDTO": {
				"$ref": "#/components/schemas/Pick_Implementation.iri-or-label-or-language-or-filePath-or-functionName-or-returnType-or-validation_",
				"description": "实现DTO - 实现信息的传输格式"
			},
...
			"ModelImplementationsResponse": {
				"description": "模型实现列表响应 (GET /api/kg/models/:id/implementations)",
				"properties": {
					"modelId": {
						"type": "string"
					},
					"modelLabel": {
						"type": "string"
					},
					"implementations": {
						"items": {
							"properties": {
								"functionName": {
									"type": "string"
								},
								"filePath": {
									"type": "string"
								},
								"language": {
									"type": "string"
								},
								"label": {
									"type": "string"
								},
								"iri": {
									"type": "string"
								}
							},
							"required": [
								"label",
								"iri"
							],
							"type": "object"
						},
						"type": "array"
					},
					"total": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"modelId",
					"modelLabel",
					"implementations",
					"total"
				],
				"type": "object",
				"additionalProperties": false
			},
			"AddModelImplementationResponse": {
				"description": "添加实现到模型响应",
				"properties": {
					"model_iri": {
						"type": "string"
					},
					"implementation_iri": {
						"type": "string"
					},
					"added_at": {
						"type": "string"
					}
				},
				"required": [
					"model_iri",
					"implementation_iri",
					"added_at"
				],
				"type": "object",
				"additionalProperties": false
			},
			"AddModelImplementationRequest": {
				"description": "添加实现到模型请求 (POST /api/kg/models/:id/implementations)",
				"properties": {
					"implementationId": {
						"type": "string"
					}
				},
				"required": [
					"implementationId"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"ImplementationsResponse": {
				"description": "实现列表响应",
				"properties": {
					"page": {
						"type": "number",
						"format": "double"
					},
					"size": {
						"type": "number",
						"format": "double"
					},
					"total": {
						"type": "number",
						"format": "double"
					},
					"totalPages": {
						"type": "number",
						"format": "double"
					},
					"result": {
						"items": {
							"$ref": "#/components/schemas/ImplementationDTO"
						},
						"type": "array"
					}
				},
				"required": [
					"result"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ImplementationDetailResponse": {
				"description": "实现详情响应",
				"properties": {
					"result": {
						"allOf": [
							{
								"$ref": "#/components/schemas/ImplementationDTO"
							},
							{
								"properties": {
									"updatedAt": {
										"type": "string"
									},
									"createdAt": {
										"type": "string"
									},
									"relatedModels": {
										"items": {
											"properties": {
												"calculationType": {
													"type": "string"
												},
												"label": {
													"type": "string"
												},
												"iri": {
													"type": "string"
												}
											},
											"required": [
												"label",
												"iri"
											],
											"type": "object"
										},
										"type": "array"
									},
									"inputParameters": {
										"type": "string"
									},
									"description": {
										"type": "string"
									}
								},
								"type": "object"
							}
						]
					}
				},
				"required": [
					"result"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CreateImplementationResponse": {
				"description": "创建实现响应",
				"properties": {
					"iri": {
						"type": "string"
					},
					"label": {
						"type": "string"
					},
					"language": {
						"type": "string"
					},
					"file_path": {
						"type": "string"
					},
					"created_at": {
						"type": "string"
					}
				},
				"required": [
					"iri",
					"label",
					"language",
					"file_path",
					"created_at"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CreateImplementationRequest": {
				"description": "创建实现请求",
				"properties": {
					"name": {
						"type": "string"
					},
					"language": {
						"type": "string"
					},
					"file_path": {
						"type": "string"
					},
					"function_name": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"input_parameters": {
						"type": "string"
					},
					"return_type": {
						"type": "string"
					},
					"validation": {
						"type": "string"
					}
				},
				"required": [
					"name",
					"language",
					"file_path"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UpdateImplementationResponse": {
				"description": "更新实现响应",
				"properties": {
					"iri": {
						"type": "string"
					},
					"label": {
						"type": "string"
					},
					"language": {
						"type": "string"
					},
					"file_path": {
						"type": "string"
					},
					"updated_at": {
						"type": "string"
					}
				},
				"required": [
					"iri",
					"label",
					"updated_at"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UpdateImplementationRequest": {
				"description": "更新实现请求",
				"properties": {
					"label": {
						"type": "string"
					},
					"language": {
						"type": "string"
					},
					"file_path": {
						"type": "string"
					},
					"function_name": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"input_parameters": {
						"type": "string"
					},
					"return_type": {
						"type": "string"
					},
					"validation": {
						"type": "string"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"DeleteImplementationResponse": {
				"description": "删除实现响应",
				"properties": {
					"iri": {
						"type": "string"
					},
					"deleted": {
						"type": "boolean"
					},
					"deleted_at": {
						"type": "string"
					}
				},
				"required": [
					"iri",
					"deleted",
					"deleted_at"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "ESG Knowledge Graph API",
		"version": "1.0.0",
		"description": "ESG Knowledge Graph platform backend API with auto-generated OpenAPI documentation",
		"license": {
			"name": "MIT"
		},
		"contact": {
			"name": "ESG Knowledge Graph API Support"
		}
	},
	"paths": {
		"/api/computation/method": {
			"get": {
				"operationId": "GetComputationMethodWithTsoa",
				"responses": {
					"200": {
						"description": "Success",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"description": "CQ5: Get metric calculation method information\nGET /api/computation/method?metric_label={metric_label}",
				"tags": [
					"Computation (Deprecated)"
				],
				"deprecated": true,
				"security": [],
				"parameters": [
					{
						"description": "Metric label",
						"in": "query",
						"name": "metric_label",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/computation/implementation": {
			"get": {
				"operationId": "GetImplementationInfoWithTsoa",
				"responses": {
					"200": {
						"description": "Success",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"description": "Get implementation file information\nGET /api/computation/implementation?implementation_label={implementation_label}",
				"tags": [
					"Computation (Deprecated)"
				],
				"deprecated": true,
				"security": [],
				"parameters": [
					{
						"description": "Implementation label",
						"in": "query",
						"name": "implementation_label",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/computation/implementations": {
			"get": {
				"operationId": "GetImplementationsByTypeWithTsoa",
				"responses": {
					"200": {
						"description": "Success",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"description": "Get all implementation information by calculation type\nGET /api/computation/implementations?calculation_type={calculation_type}",
				"tags": [
					"Computation (Deprecated)"
				],
				"deprecated": true,
				"security": [],
				"parameters": [
					{
						"description": "Calculation type",
						"in": "query",
						"name": "calculation_type",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
		"/api/kg/models/{id}/implementations": {
			"get": {
				"operationId": "GetModelImplementationsWithTsoa",
				"responses": {
					"200": {
						"description": "Success",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ModelImplementationsResponse"
								}
							}
						}
					}
				},
				"description": "Get model implementations\n\nReturns all implementations associated with this model via esg:executesWith.",
				"tags": [
					"Models"
				],
				"security": [],
				"parameters": [
					{
						"description": "Model identifier (IRI, label, or short ID)",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"post": {
				"operationId": "AddModelImplementationWithTsoa",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AddModelImplementationResponse"
								}
							}
						}
					}
				},
				"description": "Add implementation to model\n\nAssociates an implementation with the model via esg:executesWith relationship.",
				"tags": [
					"Models"
				],
				"security": [],
				"parameters": [
					{
						"description": "Model identifier (IRI, label, or short ID)",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"description": "Request body containing implementation identifier",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/AddModelImplementationRequest",
								"description": "Request body containing implementation identifier"
							}
						}
					}
				}
			}
		},
...
		"/api/kg/implementations": {
			"get": {
				"operationId": "GetImplementationsWithTsoa",
				"responses": {
					"200": {
						"description": "Success",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ImplementationsResponse"
								}
							}
						}
					}
				},
				"description": "Get implementation list (supports pagination and search)",
				"tags": [
					"Implementations"
				],
				"security": [],
				"parameters": [
					{
						"description": "Page number (starting from 1, default 1)",
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"description": "Items per page (default 10, max 100)",
						"in": "query",
						"name": "size",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"description": "Search keyword (fuzzy match on label)",
						"in": "query",
						"name": "search",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "Filter by programming language (optional)",
						"in": "query",
						"name": "language",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "Filter by file path (optional)",
						"in": "query",
						"name": "filePath",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "Filter by calculation type (via associated models) (optional)",
						"in": "query",
						"name": "calculationType",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "Sort field (default label)",
						"in": "query",
						"name": "sort",
						"required": false,
						"schema": {
							"type": "string",
							"enum": [
								"label",
								"createdAt"
							]
						}
					},
					{
						"description": "Sort order (default asc)",
						"in": "query",
						"name": "order",
						"required": false,
						"schema": {
							"type": "string",
							"enum": [
								"asc",
								"desc"
							]
						}
					}
				]
			},
			"post": {
				"operationId": "CreateImplementationWithTsoa",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CreateImplementationResponse"
								}
							}
						}
					}
				},
				"description": "Create a new implementation",
				"tags": [
					"Implementations"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"description": "Request data for creating an implementation",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateImplementationRequest",
								"description": "Request data for creating an implementation"
							}
						}
					}
				}
			}
		},
		"/api/kg/implementations/{id}": {
			"get": {
				"operationId": "GetImplementationByIdWithTsoa",
				"responses": {
					"200": {
						"description": "Success",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ImplementationDetailResponse"
								}
							}
						}
					}
				},
				"description": "Get implementation details",
				"tags": [
					"Implementations"
				],
				"security": [],
				"parameters": [
					{
						"description": "Implementation ID (can be URI, namespace format or short ID)",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						},
						"example": "Implementation_Python_PercentageRatio"
					}
				]
			},
			"patch": {
				"operationId": "UpdateImplementationWithTsoa",
				"responses": {
					"200": {
						"description": "Success",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UpdateImplementationResponse"
								}
							}
						}
					}
				},
				"description": "Partial update of implementation",
				"tags": [
					"Implementations"
				],
				"security": [],
				"parameters": [
					{
						"description": "Implementation ID",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"description": "Request data for updating the implementation",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateImplementationRequest",
								"description": "Request data for updating the implementation"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteImplementationWithTsoa",
				"responses": {
					"200": {
						"description": "Success",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/DeleteImplementationResponse"
								}
							}
						}
					}
				},
				"description": "Delete an implementation",
				"tags": [
					"Implementations"
				],
				"security": [],
				"parameters": [
					{
						"description": "Implementation ID",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "Whether to force delete (ignore model reference association checks)",
						"in": "query",
						"name": "force",
						"required": false,
						"schema": {
							"type": "boolean"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "https://[object Object]/"
		},
		{
			"url": "https://[object Object]/"
		}
	],
...
}